move customer to separate folder
