+++ public/src/components/Manager/Customers/DetailView.vue
