+++ public/src/router.js
@@ -8,8 +8,8 @@ import AddMaterial from '@/components/Manager/AddMaterial'
 import Login from '@/components/Manager/Login/Login'
 import WorkerTimes from '@/components/Manager/WorkerTimes/Form'
 import Customers from '@/components/Manager/Customers'
-import NewCustomer from '@/components/Manager/NewCustomer'
-import CustomerDetailView from '@/components/Manager/CustomerDetailView'
+import NewCustomer from '@/components/Manager/Customers/NewCustomer'
+import CustomerDetailView from '@/components/Manager/Customers/DetailView'
 import Menu from '@/components/Manager/Menu/Menu'
 import AddStock from '@/components/Manager/AddStock/AddStock'
 import EditStockForm from '@/components/Manager/AddStock/EditStockForm'
